+++ lib/MqttClient.ts
@@ -1,13 +1,14 @@
 'use strict'
 
-// eslint-disable-next-line one-var
 import mqtt, { Client } from 'mqtt'
 import { allSettled, parseJSON, sanitizeTopic } from './utils'
-// import { storeDir } from '../config/app'
 import { module } from './logger'
 import { version as appVersion } from '../package.json'
 import { TypedEventEmitter } from './EventEmitter'
-// import LevelStore from 'mqtt-level-store'
+import { storeDir } from '../config/app'
+import { ensureDir } from 'fs-extra'
+import { Manager } from 'mqtt-jsonl-store'
+import { join } from 'path'
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const url = require('native-url')
@@ -86,7 +87,9 @@ class MqttClient extends TypedEventEmitter<MqttClientEventCallbacks> {
 	 */
 	constructor(config: MqttConfig) {
 		super()
-		this._init(config)
+		this._init(config).catch((e) => {
+			logger.error('Error while initializing MQTT Client', e)
+		})
 	}
 
 	get connected() {
@@ -203,7 +206,7 @@ class MqttClient extends TypedEventEmitter<MqttClientEventCallbacks> {
 
 		logger.info('Restarting Mqtt Client after update...')
 
-		this._init(config)
+		await this._init(config)
 	}
 
 	/**
@@ -315,7 +318,7 @@ class MqttClient extends TypedEventEmitter<MqttClientEventCallbacks> {
 	/**
 	 * Initialize client
 	 */
-	private _init(config: MqttConfig) {
+	private async _init(config: MqttConfig) {
 		this.config = config
 		this.toSubscribe = new Map()
 
@@ -357,12 +360,16 @@ class MqttClient extends TypedEventEmitter<MqttClientEventCallbacks> {
 			}
 		}
 
-		// Temporary fix, seems mqtt store prevents mqtt to work...
-		// if (config.store) {
-		// 	const manager = LevelStore(joinPath(storeDir, 'mqtt'))
-		// 	options.incomingStore = manager.incoming
-		// 	options.outgoingStore = manager.outgoing
-		// }
+		if (config.store) {
+			const dbDir = join(storeDir, 'mqtt-packets-store')
+			await ensureDir(dbDir)
+			const manager = new Manager(dbDir)
+			await manager.open()
+
+			// no reason to use a memory store for incoming messages
+			options.incomingStore = manager.incoming
+			options.outgoingStore = manager.outgoing
+		}
 
 		if (config.auth) {
 			options.username = config.username
@@ -405,7 +412,6 @@ class MqttClient extends TypedEventEmitter<MqttClientEventCallbacks> {
 		)
 
 		// subscribe to actions
-		// eslint-disable-next-line no-redeclare
 		for (let i = 0; i < MqttClient.ACTIONS.length; i++) {
 			subscribePromises.push(
 				this.subscribe(
@@ -498,8 +504,9 @@ class MqttClient extends TypedEventEmitter<MqttClientEventCallbacks> {
 		if (isNaN(parseInt(parsed))) {
 			try {
 				parsed = parseJSON(parsed)
-				// eslint-disable-next-line no-empty
-			} catch (e) {} // it' ok fallback to string
+			} catch (e) {
+				// it' ok fallback to string
+			}
 		} else {
 			parsed = Number(parsed)
 		}